requirements.txt
# Core dependencies
binance
python-dotenv
